+++ src/tools/rect.c
@@ -6,7 +6,7 @@
 /*   By: sbecker <sbecker@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/10 15:52:59 by sbednar           #+#    #+#             */
-/*   Updated: 2019/07/13 10:09:37 by sbecker          ###   ########.fr       */
+/*   Updated: 2019/07/15 12:59:30 by sbecker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,8 +38,8 @@ void	draw_empty_rect(t_guimp *g, t_vec2 s, t_vec2 e)
 	int py = ft_min(s.y, e.y);
 	int cx = s.x + e.x - px;
 	int cy = s.y + e.y - py;
-	if (abs(s.x - e.x) < g->draw_tool.brush_size ||
-		abs(s.y - e.y) < g->draw_tool.brush_size)
+	if ((Uint32)abs(s.x - e.x) < g->draw_tool.brush_size ||
+		(Uint32)abs(s.y - e.y) < g->draw_tool.brush_size)
 	{
 		draw_rect(g, s, e);
 		return ;
